requirements.txt
bs4
lxml
selenium
requests
sqlalchemy
python-decouple
webdriver-manager
fake-useragent
pandas
psycopg2-binary